Rename alfasim_sdk.data_type and alfasim_sdk.data_models to types and models
